+++ CHANGELOG.md
@@ -1,8 +1,19 @@
-# 1.0.1+1
+## 1.1.0+1 - 2023-04-30
+
+### Added
+
+- Support to turn a logger on or of with `shouldLog`.
+- New `EmojiPrinter` for logging with emojis 🚀
+
+### Fixed
+
+- spelling mistakes
+
+## 1.0.1+1
 
 - **fix:** changed example folder
 - **doc:** made package description longer
 
-# 1.0.0+1
+## 1.0.0+1
 
 - **feat:** initial commit 🎉